Move loop_selection
